[refactor] order query module 분리
