--- order-api/order-application/src/main/java/org/flab/deliveryplatform/order/application/port/GetMyOrdersQuery.java
@@ -1,9 +1,0 @@
-package org.flab.deliveryplatform.order.application.port;
-
-import java.util.List;
-import org.flab.deliveryplatform.order.application.port.dto.OrderData;
-
-public interface GetMyOrdersQuery {
-
-    List<OrderData> getMyOrders(Long memberId);
-}
