--- order-api/order-application/src/main/java/org/flab/deliveryplatform/order/application/port/OrderQueryRepository.java
@@ -1,9 +1,0 @@
-package org.flab.deliveryplatform.order.application.port;
-
-import java.util.List;
-import org.flab.deliveryplatform.order.application.port.dto.OrderData;
-
-public interface OrderQueryRepository {
-
-    List<OrderData> findAllByMemberId(Long memberId);
-}
