+++ order-query-api/order-query-application/build.gradle
@@ -1,0 +1,7 @@
+version '1.0-SNAPSHOT'
+
+dependencies {
+    implementation project(':common')
+    implementation 'org.springframework.boot:spring-boot-starter'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+}
