+++ order-query-api/order-query-application/src/main/java/org/flab/deliveryplatform/order/query/application/OrderQueryRepository.java
@@ -1,0 +1,9 @@
+package org.flab.deliveryplatform.order.query.application;
+
+import java.util.List;
+import org.flab.deliveryplatform.order.query.application.dto.OrderData;
+
+public interface OrderQueryRepository {
+
+    List<OrderData> findAllByMemberId(Long memberId);
+}
