+++ order-query-api/order-query-application/src/main/java/org/flab/deliveryplatform/order/query/application/service/GetMyOrdersService.java
@@ -1,10 +1,10 @@
-package org.flab.deliveryplatform.order.application.service;
+package org.flab.deliveryplatform.order.query.application.service;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.flab.deliveryplatform.order.application.port.GetMyOrdersQuery;
-import org.flab.deliveryplatform.order.application.port.OrderQueryRepository;
-import org.flab.deliveryplatform.order.application.port.dto.OrderData;
+import org.flab.deliveryplatform.order.query.application.GetMyOrdersQuery;
+import org.flab.deliveryplatform.order.query.application.OrderQueryRepository;
+import org.flab.deliveryplatform.order.query.application.dto.OrderData;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
