+++ order-query-api/order-query-interfaces/build.gradle
@@ -1,0 +1,7 @@
+version '1.0-SNAPSHOT'
+
+dependencies {
+    implementation project(':order-query-api:order-query-application')
+    implementation project(':common')
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+}
