+++ settings.gradle
@@ -47,4 +47,10 @@ include 'delivery-api:delivery-domain'
 findProject(':delivery-api:delivery-domain')?.name = 'delivery-domain'
 include 'delivery-api:delivery-infrastructure'
 findProject(':delivery-api:delivery-infrastructure')?.name = 'delivery-infrastructure'
-
+include 'order-query-api'
+include 'order-query-api:order-query-interfaces'
+findProject(':order-query-api:order-query-interfaces')?.name = 'order-query-interfaces'
+include 'order-query-api:order-query-application'
+findProject(':order-query-api:order-query-application')?.name = 'order-query-application'
+include 'order-query-api:order-query-infrastructure'
+findProject(':order-query-api:order-query-infrastructure')?.name = 'order-query-infrastructure'
